chore: move tests to proper folder
