+++ packages/appkit/tests/mocks/Adapter.ts
@@ -1,5 +1,5 @@
 import { vi } from 'vitest'
-import type { AdapterBlueprint } from '../../adapters/ChainAdapterBlueprint.js'
+import type { AdapterBlueprint } from '../../src/adapters/ChainAdapterBlueprint.js'
 
 export const mockUniversalAdapter = {
   namespace: 'eip155',
