+++ packages/appkit/tests/mocks/Options.ts
@@ -1,6 +1,6 @@
 import type { ChainAdapter } from '@reown/appkit-core'
-import type { AppKitOptions } from '../../utils/index.js'
-import { mainnet, solana } from '../../networks/index.js'
+import type { AppKitOptions } from '../../src/utils/index.js'
+import { mainnet, solana } from '../../src/networks/index.js'
 import type { SdkVersion } from '@reown/appkit-core'
 import { vi } from 'vitest'
